requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
distlib==0.3.9
Django==5.1.7
djangorestframework==3.15.2
drf-spectacular==0.28.0
filelock==3.18.0
flake8==7.1.2
identify==2.6.9
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mccabe==0.7.0
nodeenv==1.9.1
platformdirs==4.3.6
pre_commit==4.1.0
pycodestyle==2.12.1
pyflakes==3.2.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
sqlparse==0.5.3
stripe==11.6.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.29.3
